packages/mantle/src/components/text-area/text-area.test.tsx
import { render, screen } from "@testing-library/react";
import { describe, expect, test } from "vitest";
import { TextArea } from "./text-area.js";

describe("TextArea", () => {
	test('given validation={false}, renders a textarea with aria-invalid="false" and not have data-validation', () => {
		render(<TextArea validation={false} />);
		expect(screen.getByRole("textbox")).toHaveAttribute("aria-invalid", "false");
		expect(screen.getByRole("textbox")).not.toHaveAttribute("data-validation");
	});

	test('given validation="success", renders a textarea with aria-invalid="false" and data-validation="success"', () => {
		render(<TextArea validation="success" />);
		expect(screen.getByRole("textbox")).toHaveAttribute("aria-invalid", "false");
		expect(screen.getByRole("textbox")).toHaveAttribute("data-validation", "success");
	});

	test('given validation="warning", renders a textarea with aria-invalid="false" and data-validation="warning"', () => {
		render(<TextArea validation="warning" />);
		expect(screen.getByRole("textbox")).toHaveAttribute("aria-invalid", "false");
		expect(screen.getByRole("textbox")).toHaveAttribute("data-validation", "warning");
	});

	test('given validation="error", renders a textarea with aria-invalid="true" and data-validation="error"', () => {
		render(<TextArea validation="error" />);
		expect(screen.getByRole("textbox")).toHaveAttribute("aria-invalid", "true");
		expect(screen.getByRole("textbox")).toHaveAttribute("data-validation", "error");
	});

	test('given aria-invalid="true" and validation="success", renders a textarea with aria-invalid="true" and data-validation="error"', () => {
		render(<TextArea aria-invalid="true" validation="success" />);
		expect(screen.getByRole("textbox")).toHaveAttribute("aria-invalid", "true");
		expect(screen.getByRole("textbox")).toHaveAttribute("data-validation", "error");
	});

	test('given aria-invalid="true" and validation="warning", renders a textarea with aria-invalid="true" and data-validation="error"', () => {
		render(<TextArea aria-invalid="true" validation="warning" />);
		expect(screen.getByRole("textbox")).toHaveAttribute("aria-invalid", "true");
		expect(screen.getByRole("textbox")).toHaveAttribute("data-validation", "error");
	});

	test('given aria-invalid="true" and validation="error", renders a textarea with aria-invalid="true" and data-validation="error"', () => {
		render(<TextArea aria-invalid="true" validation="error" />);
		expect(screen.getByRole("textbox")).toHaveAttribute("aria-invalid", "true");
		expect(screen.getByRole("textbox")).toHaveAttribute("data-validation", "error");
	});
});
